DriveSystemAutonomous added and remove autonomous methods from normal drivesystem
