+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/components/ArmSystem.java
@@ -260,7 +260,7 @@ public class ArmSystem {
     public boolean isGettingCapstone() { return gettingCapstone; }
 
     // Moves slider back to original state
-    public void stop() {
-        slider.setTargetPosition(calibrationDistance);
-    }
+//    public void stop() {
+//        slider.setTargetPosition(calibrationDistance);
+//    }
 }
